requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.4
faiss-cpu==1.13.2
idna==3.11
numpy==2.4.2
packaging==26.0
requests==2.32.5
urllib3==2.6.3
